On-Behalf_Of consenting implemented using the supported Microsoft Identity Platform for SSB
Implements an incremental consenting approach.
Currently the sample app implements the auth-start and auth-end and (potentially) error endpoints. Consider how this can be moved to the library and still allow customisation.
Consider reverting the changes to UserInfoView and create a separate demo page - up to 16 personal tabs can be included in a teams integration so this could be a worthwhile test.